requirements.txt
streamlit==0.55.0
pandas==1.0.1
numpy==1.18.1
scikit-learn==0.22.1
matplotlib==3.1.3
eli5==0.10.0
pylint==2.4.4
shap==0.35.0
plotly==4.6.0
Pillow==7.1.1